add getAllNotes

+++ src/api/notes/handler.js
@@ -10,13 +10,13 @@ class NotesHandler {
 
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
-
     const {
       title, cue, main, summary, categoryId,
     } = request.payload;
+    const { id: credentialId } = request.auth.credentials;
 
     const noteId = await this._service.addNote({
-      title, cue, main, summary, categoryId,
+      title, cue, main, summary, categoryId, owner: credentialId,
     });
 
     const response = h.response({
@@ -30,8 +30,22 @@ class NotesHandler {
     return response;
   }
 
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
+    return {
+      status: 'success',
+      data: {
+        notes,
+      },
+    };
+  }
+
   async getNoteByIdHandler(request) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyOwner(id, credentialId);
+
     const note = await this._service.getNoteById(id);
     return {
       status: 'success',
@@ -44,7 +58,9 @@ class NotesHandler {
   async putNoteByIdHandler(request) {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
 
+    await this._service.verifyOwner(id, credentialId);
     await this._service.editNoteById(id, request.payload);
 
     return {
@@ -55,6 +71,9 @@ class NotesHandler {
 
   async deleteNoteByIdHandler(request) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+
+    await this._service.verifyOwner(id, credentialId);
     await this._service.deleteNoteById(id);
     return {
       status: 'success',
